// src/app/page/master/messaging/messaging.interface.ts
import {EventEmitter} from '@angular/core';

export interface Messaging {
	kdMessaging?: any;
	namaMessaging?: string;
	reportDisplay?: string;
	kdMessagingHead?: string;
	namaMessagingHead?: string;
	deskripsiDetailMessaging?: string;
	kdObjekModulAplikasi?: string;
	keteranganLainnya?:string;
	statusEnabled?: boolean;
	noRec?:string;
	id?:any;
  	kode?:any;
  	version?:any;
  	label?: any;
  	value?: any;
  	kodeExternal?: string;
    namaExternal?: string;
 }